Decouple Quantizer from AnalysisPredictor
test=develop
